Make clang-tidy header checks self-contained
